+++ experimental/packages/opentelemetry-sdk-node/test/sdk.test.ts
@@ -18,6 +18,7 @@ import {
   context,
   propagation,
   ProxyTracerProvider,
+  ProxyMeterProvider,
   trace,
   diag,
   DiagLogLevel,
@@ -90,6 +91,7 @@ describe('Node SDK', () => {
   let ctxManager: any;
   let propagator: any;
   let delegate: any;
+  let metricsDelegate: any;
   let logsDelegate: any;
 
   beforeEach(() => {
@@ -103,6 +105,9 @@ describe('Node SDK', () => {
     ctxManager = context['_getContextManager']();
     propagator = propagation['_getGlobalPropagator']();
     delegate = (trace.getTracerProvider() as ProxyTracerProvider).getDelegate();
+    metricsDelegate = (
+      metrics.getMeterProvider() as ProxyMeterProvider
+    ).getDelegate();
     logsDelegate = (
       logs.getLoggerProvider() as ProxyLoggerProvider
     ).getDelegate();
@@ -140,6 +145,12 @@ describe('Node SDK', () => {
         delegate,
         'tracer provider should not have changed'
       );
+      assert.strictEqual(
+        (metrics.getMeterProvider() as ProxyMeterProvider).getDelegate(),
+        metricsDelegate,
+        'meter provider should not have changed'
+      );
+      assert.ok(!(logs.getLoggerProvider() instanceof LoggerProvider));
       assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
       assert.strictEqual(
         (logs.getLoggerProvider() as ProxyLoggerProvider).getDelegate(),
@@ -193,7 +204,13 @@ describe('Node SDK', () => {
 
       sdk.start();
 
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
+      assert.ok(
+        !(
+          (
+            metrics.getMeterProvider() as ProxyMeterProvider
+          ).getDelegate() instanceof MeterProvider
+        )
+      );
 
       assert.ok(
         context['_getContextManager']().constructor.name ===
@@ -216,7 +233,13 @@ describe('Node SDK', () => {
 
       sdk.start();
 
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
+      assert.ok(
+        !(
+          (
+            metrics.getMeterProvider() as ProxyMeterProvider
+          ).getDelegate() instanceof MeterProvider
+        )
+      );
 
       assert.ok(
         context['_getContextManager']().constructor.name ===
@@ -246,7 +269,13 @@ describe('Node SDK', () => {
 
       sdk.start();
 
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
+      assert.ok(
+        !(
+          (
+            metrics.getMeterProvider() as ProxyMeterProvider
+          ).getDelegate() instanceof MeterProvider
+        )
+      );
 
       assert.ok(
         context['_getContextManager']().constructor.name ===
@@ -315,9 +344,11 @@ describe('Node SDK', () => {
         delegate,
         'tracer provider should not have changed'
       );
-
-      assert.ok(metrics.getMeterProvider() instanceof MeterProvider);
-
+      assert.ok(
+        (
+          metrics.getMeterProvider() as ProxyMeterProvider
+        ).getDelegate() instanceof MeterProvider
+      );
       await sdk.shutdown();
       delete env.OTEL_TRACES_EXPORTER;
     });
@@ -488,8 +519,8 @@ describe('Node SDK', () => {
       'tracer provider should not have changed'
     );
 
-    const meterProvider = metrics.getMeterProvider() as MeterProvider;
-    assert.ok(meterProvider);
+    const meterProvider = metrics.getMeterProvider() as ProxyMeterProvider;
+    assert.ok(meterProvider.getDelegate() instanceof MeterProvider);
 
     const meter = meterProvider.getMeter('NodeSDKViews', '1.0.0');
     const counter = meter.createCounter('test_counter', {
@@ -882,7 +913,13 @@ describe('Node SDK', () => {
       });
       sdk.start();
 
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
+      assert.ok(
+        !(
+          (
+            metrics.getMeterProvider() as ProxyMeterProvider
+          ).getDelegate() instanceof MeterProvider
+        )
+      );
 
       await sdk.shutdown();
     });
@@ -1127,7 +1164,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           ConsoleMetricExporter
@@ -1149,7 +1188,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPGrpcMetricExporter
@@ -1171,7 +1212,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPProtoMetricExporter
@@ -1193,7 +1236,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPHttpMetricExporter
@@ -1215,7 +1260,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPGrpcMetricExporter
@@ -1237,7 +1284,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPProtoMetricExporter
@@ -1259,7 +1308,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPProtoMetricExporter
@@ -1284,7 +1335,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader._exporter instanceof
           OTLPProtoMetricExporter
@@ -1308,7 +1361,9 @@ describe('Node SDK', () => {
       const sdk = new NodeSDK();
       sdk.start();
       const meterProvider = metrics.getMeterProvider();
-      const sharedState = (meterProvider as any)['_sharedState'];
+      const sharedState = (
+        (meterProvider as ProxyMeterProvider).getDelegate() as any
+      )['_sharedState'];
       assert.ok(
         sharedState.metricCollectors[0]._metricReader instanceof
           PrometheusMetricExporter